Convert to Python package
This commit changes this project into a Python package, as well
as goes through renaming things to make the project configuration
and commands use `throttled` instead of `lenovo_fix`.  It also
takes care of setting up data_files so that the /etc configuration
files are set up out of the box simply by installing the Python
package.

This greatly simplifies the process of potentially packaging this
project as well as makes it easy to simply pip install it into your
system.  It also loosens up the requirements which means that it's
likely to easily install using python packages shipped by distros.